Third value in ServerInfo is a uint64, not a float.

+++ RoamingSettings.cpp
@@ -15,7 +15,7 @@ void RoamingSettings::parse(Json::Value root) {
                     ServerInfo info;
                     info.serverRegion = root[2][it.index()][0].asUInt();
                     info.unkownVal1 = root[2][it.index()][1].asUInt();
-                    info.unkownVal2 = root[2][it.index()][2].asFloat();
+                    info.unkownVal2 = root[2][it.index()][2].asUInt64();
                     info.regionCode = root[2][it.index()][3].asString();
                     servers.push_back(info);
                 }
